+++ src/Control/Lens/Fold.hs
@@ -38,7 +38,8 @@ module Control.Lens.Fold
   (
   -- * Folds
     Fold
-  , (^?), (^..)
+  , (^..)
+  , (^?) , (^?!)
   -- ** Building Folds
   --, folds
   , folding
@@ -91,7 +92,7 @@ import Data.Monoid
 -- $setup
 -- >>> import Control.Lens
 
-infixl 8 ^?, ^..
+infixl 8 ^.., ^?, ^?!
 
 --------------------------
 -- Folds
@@ -744,6 +745,18 @@ headOf l = getFirst# (foldMapOf l (first# Just))
 (^?) :: s -> Getting (First a) s t a b -> Maybe a
 a ^? l = getFirst (foldMapOf l (first# Just) a)
 {-# INLINE (^?) #-}
+
+-- | Perform an *UNSAFE* 'head' of a 'Fold' or 'Control.Lens.Traversal.Traversal' assuming that it is there.
+--
+-- @
+-- ('^?') :: s -> 'Getter' s a           -> a
+-- ('^?') :: s -> 'Fold' s a             -> a
+-- ('^?') :: s -> 'Simple' 'Lens' s a      -> a
+-- ('^?') :: s -> 'Simple' 'Control.Lens.Iso.Iso' s a       -> a
+-- ('^?') :: s -> 'Simple' 'Control.Lens.Traversal.Traversal' s a -> a
+-- @
+(^?!) :: s -> Getting (First a) s t a b -> a
+a ^?! l = fromMaybe (error "(^?!): empty Fold") $ getFirst (foldMapOf l (first# Just) a)
 
 -- | Perform a safe 'last' of a 'Fold' or 'Control.Lens.Traversal.Traversal' or retrieve 'Just' the result
 -- from a 'Getter' or 'Lens'.